+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fix NPE on skinupdate bungeecord
 * Fix missing skinupdate < other > command for bungee
 * Fix json parsing from third party api
+* Add BungeeCord support for Sponge servers
+* Add missing skin permissions for the Sponge server
 
 ##### 1.9